both classes in one class
